requirements.txt
numpy
Pillow
PyYAML
ultralytics
onnx
onnxsim
onnxruntime
onnxruntime-gpu